+++ com.unity.formats.fbx/Editor/FbxExporter.cs
@@ -698,8 +698,8 @@ namespace UnityEditor.Formats.Fbx.Exporter
                 if (!ExportSettings.DisableAbsolutePathWarningProperty)
                     Debug.LogWarning(
                         "The source object contains a reference to a texture file. As a consequence, the exported FBX will contain " +
-                        "an absolute reference to this texture file.\nIf you plan to share this FBX asset, double check your path does " +
-                        "not contain sensitive information.\nThis warning can be disabled in the Fbx Export section of your Project Settings." +
+                        "an absolute reference to this texture file.\nIf you plan to share this FBX asset, make sure your path does " +
+                        "not contain sensitive information.\nYou can disable this warning in the Fbx Export section of your Project Settings." +
                         $"\nTexture path: {textureSourceFullPath}");
             }
             fbxTexture.ConnectDstProperty(fbxMaterialProperty);