+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 - Fixed an exception occurring during hierarchy export when an LOD contains a null renderer.
+- Removed the deprecated Component Updater from FBX Export Project Settings.
 
 ## [5.1.1] - 2024-03-18
 ### Changed